Renamed opcodes module; removed common_types module
