--- rustzx-z80/src/common_types.rs
@@ -1,52 +1,0 @@
-//! Collection of independent types, used by other types of z80 module
-
-/// Instruction prefix type
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub enum Prefix {
-    None,
-    CB,
-    DD,
-    ED,
-    FD,
-}
-impl Prefix {
-    /// Returns prefix type from byte value
-    pub fn from_byte(data: u8) -> Prefix {
-        match data {
-            0xCB => Prefix::CB,
-            0xDD => Prefix::DD,
-            0xED => Prefix::ED,
-            0xFD => Prefix::FD,
-            _ => Prefix::None,
-        }
-    }
-
-    /// Transforms prefix back to byte
-    pub fn to_byte(self) -> Option<u8> {
-        match self {
-            Prefix::DD => Some(0xDD),
-            Prefix::FD => Some(0xFD),
-            Prefix::ED => Some(0xED),
-            Prefix::CB => Some(0xCB),
-            Prefix::None => None,
-        }
-    }
-}
-
-/// Interrupt mode enum
-#[derive(Debug, Clone, Copy)]
-pub enum IntMode {
-    IM0,
-    IM1,
-    IM2,
-}
-
-impl From<IntMode> for u8 {
-    fn from(mode: IntMode) -> Self {
-        match mode {
-            IntMode::IM0 => 0,
-            IntMode::IM1 => 1,
-            IntMode::IM2 => 2,
-        }
-    }
-}
