+++ rustzx-z80/src/opcode/internal_rot.rs
@@ -1,5 +1,5 @@
 use crate::{
-    opcodes::BitOperand8, smallnum::U3, tables::SZPF3F5_TABLE, Flag, Z80Bus, FLAG_CARRY, Z80,
+    opcode::BitOperand8, smallnum::U3, tables::SZPF3F5_TABLE, Flag, Z80Bus, FLAG_CARRY, Z80,
 };
 
 /// Rotate operations (RLC, RRC, RL, RR, SLA, SRA, SLL, SRL)
