+++ rustzx-z80/src/opcode/types.rs
@@ -3,6 +3,40 @@ use crate::{
     RegName8,
 };
 
+/// Instruction prefix type
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+pub enum Prefix {
+    None,
+    CB,
+    DD,
+    ED,
+    FD,
+}
+
+impl Prefix {
+    /// Returns prefix type from byte value
+    pub fn from_byte(data: u8) -> Prefix {
+        match data {
+            0xCB => Prefix::CB,
+            0xDD => Prefix::DD,
+            0xED => Prefix::ED,
+            0xFD => Prefix::FD,
+            _ => Prefix::None,
+        }
+    }
+
+    /// Transforms prefix back to byte
+    pub fn to_byte(self) -> Option<u8> {
+        match self {
+            Prefix::DD => Some(0xDD),
+            Prefix::FD => Some(0xFD),
+            Prefix::ED => Some(0xED),
+            Prefix::CB => Some(0xCB),
+            Prefix::None => None,
+        }
+    }
+}
+
 /// Operand for 8-bit LD instructions
 pub enum LoadOperand8 {
     Indirect(u16),
